third_party/blink/web_tests/platform/generic/http/tests/xmlhttprequest/filename-encoding-expected.txt

Verify that attachment filenames are not normalized

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS file.name is request.filename
PASS response.field is request.field
PASS response.filename is request.filename
PASS response.type is request.type
PASS response.content is request.content
PASS successfullyParsed is true

TEST COMPLETE
